feat(make-bmp): seperate logic and example
