+++ lib/bmp.js
@@ -4,8 +4,6 @@ const {
   calculatePaddingSize,
 } = require('./common/bmp');
 const { bytesPerSegment } = require('./common/constants/bytes');
-const fs = require('fs');
-
 const getFullBmp = ({ colorsData, width, height }) => {
   const fileHeader = createFileHeader({ width, height });
   const fileInfoHeader = createFileInfoHeader({ width, height });
@@ -28,19 +26,6 @@ const getFullBmp = ({ colorsData, width, height }) => {
   }
   return Buffer.concat([fileHeader, fileInfoHeader, imageData]);
 };
-
-const redBlock = Buffer.from([0, 0, 255]);
-const greenBlock = Buffer.from([0, 255, 0]);
-const blueBlock = Buffer.from([255, 0, 0]);
-
-const redBlueBlock = Buffer.concat([redBlock, redBlock, blueBlock, blueBlock]);
-
-const myBmp = getFullBmp({ colorsData: redBlueBlock, width: 2, height: 2 });
-
-fs.writeFile('test.bmp', myBmp, (err) => {
-  if (err) throw err;
-  console.log('이미지가 생성되었습니다. "test.bmp" 파일을 확인해보세요.');
-});
 
 module.exports = {
   getFullBmp,
