+++ lib/common/bmp.js
