+++ lib/common/constants/bytes.js
